Updated openstack/openstack Project: openstack/cinder  f00e7e39844b2c4ca1207525a0429c44ba335d02
Make PureISCSIDriver iSCSI port discovery more flexible

While searching for reachable iSCSI ports on the target flash array it
might fail the discovery command. If this happens on driver setup it
will throw an exception from attempting to use fields on an
uninitialized class variable. This fixes the log message and adds a
retry to the method to make this more flexible with network timeouts.

Closes-Bug: 1417730
Change-Id: I99ffb8de4dd5a421442df1bf745808505775480a