Updated openstack/openstack Project: openstack/cinder  2a1705564afee7be7c77c1e1d3e8844b1a207b66
Create SolidFire Template account on init

The standard flow in the SolidFire volume is to create
accounts dynamically as needed based on project id.
Unfortunately the clone_image method "forgot" about this
detail and calls the private clone_volume method without
a valid account resulting in an AccountNotFound
Exception.

The clone_image method is enabled using a specific account
provided via the configuration file.  Rather than worry
about dynamic account creation on clone_image call, this
patch just makes it part of the init phase.  We check to
see if template caching is enabled and if it is, just
go ahead and check/create the account on startup.

Change-Id: I41a8d400d3dfc0e185304f80efd419ebac9efdf6
Closes-Bug: #1414760