Updated openstack/openstack Project: openstack/nova  ff219370f3e1c836603b218dd3225f81ba3938b4
Fix references to non-existent "pause" section

Two sections of the hypervisor support matrix refer readers to the
sections of the document dealing with the "pause" operation.
Although there is a section in the raw .ini file called
"[operation.pause]", this doesn't appear at all in the rendered
version of document since it's title is set to
"Stop instance CPUs".  This renders readers of the rendered
document unable to find sections being referenced.

This patch clears up the confusion by changing the references
to refer to "Stop instance CPUs" rather than "pause".

Change-Id: Ia97d8e400db3a11291a715f6a630ea8bc638993e
Closes-bug: #1418641