Updated openstack/openstack Project: openstack/nova  bad4af5d217b6393a7fda8f5e7b36c8fe8872667
Fix bdm transformation for volume backed servers

When volume backed server is snapshoted with nova image-create
then the image is created with a proper bdm. However when this
image is booted with image->volume bdm the two bdms are not
merged properly and the server cannot be booted as it gets two
boot_index = 0 device

This patch solves the issue by removing the unnecessary device
from the mapping.

Closes-bug: #1381598

Change-Id: I170e80d35fe8a8d906370a507148351926763a13