Updated openstack/openstack Project: openstack/nova  69f5eb1bc567595f4f1f3ffb43ce50be004fd058
Switch off oslo.* namespace check temporarily

There's lot of code in flight that will break if we switch
on the asserts to error out pep8/flake8 right now. So
we should temporarily switch it off. Doing this as a separate
review so it's easy to revery and stands out on its own.

Change-Id: I7033847bab82b6d3cd8f2bdb66c553fec93d7acb