Updated openstack/openstack Project: openstack/tempest  962ce07856b99f1611f86e2fa7371f8bf40bece3
Fix import in nova cli tests

The min version decorator was references the full qualified name for
the decorator but the cli module is imported as just cli in the test
module. Why this isn't causing an error remains unknown (it fails
flake8 locally on my system) but regardless this patch fixes the
issue and corrects the call.

Change-Id: I916d6621ab93900355fa1cb2b114afecbdc60b33