Updated openstack/openstack Project: openstack/python-novaclient  578390ee7e74c0998b9590f0327449ae19a5c8f0
Change the unsuitable item caching for completion

This fix changes the item that cached for completion from
"human_id" to "name".
A string for completion shoud not be changed in any way,
because changed keyword can not use as command line parameter.
But the "human_id" means "human readable id" that is
changed from "name" by method "to_slug". "to_slug" is meant to
create a valid path name from a string.

The tools/nova.bash_completion take the completion string from
the file ".novaclient/*/*-cache". The file is created when client
call "list" command. For example , "nova list","nova image-list",
and the others.
Currently, items that are written to the cache file is the "id"
and "human_id".

Closes-Bug: #1193049

Change-Id: I241ec8b7c8729274ee43db6e360141fd381b265e