Updated openstack/openstack Project: openstack/python-keystoneclient  6106b2be8393e6cb4256b5a2916f065e67e673f9
Add name parameter to NoMatchingPlugin exception

If you want to handle a NoMatchingPlugin exception rather than simply
exit then the name of the missing plugin is generally more useful than
the message.

The exception is specific enough that you can know what went wrong, but
you cannot determine the name of the missing plugin if you want to do
your own logging - only use the message that is generated.

We should keep the message but expose the plugin name as well.

Closes-Bug: #1410391
Change-Id: Ic93ec6583b8d7797529d36d63995ef0d8db754f1