Updated openstack/openstack Project: openstack/python-keystoneclient  14aa7fe27b0dcae5d5b4be6f9c6d5948c7704aae
Basic AccessInfo plugin

Generally we want people to use the existing plugins to manage their
authentication, however there are a number of existing services that
know how to work with an AccessInfo object directly and either cache it
or manipulate it manually.

Provide a simple Identity plugin that just takes an existing AccessInfo
and allows it to be used as an authentication plugin.

Change-Id: I388283c03a0a8a3d1afe43138eebbe5e66ca9102