Updated openstack/openstack Project: openstack/ceilometer  252744e9027af413b6a751ea7de62126f241d4e0
Add local node resource for IPMI pollsters

Each pollster should have some resources for polling, but IPMI
pollster is an exception that talk with local IPMI HW directly by
ipmitool. Local node resource is faked resource that stands for
endpoint of localhost. With it, IPMI pollsters can be handled in the
same way as others.

Change-Id: Icda81ffa2334b764e7b508f26efd031f2a863e64