Updated openstack/openstack Project: openstack/python-novaclient  2504f39ce3e13dffdb70481e7c9f59f1432caad9
Wrong help for nova usage command

nova help usage commands provides help that the command
can be used with either tenant name or tenant ID but using
tenant name gives wrong output.

So, updated the help for the command so that it is shown
that only tenant ID can be used.

Change-Id: I2ae1edd28d75fc7988fcd0da5b0fdd8081455b16
Closes-Bug: #1419726