Updated openstack/openstack Project: openstack/oslo.middleware  c2011714bf9aad7952909613e64613906d5f04e8
Adds missing files

Under certain circonstances which I couldn't fully understand
(eg: it works if in a git repo, but not if the .git folder
isn't there), python setup.py install forgets some files, and
not all is there. This patch fixed all issues for me in my
Debian package.

Change-Id: I814c4e91156434b1897eb9f50af4d0b81543ca2b