Updated openstack/openstack Project: openstack/nova  9203f89191f405a5c3a297482bf93112dbff1a5a
Add VirtCPUModel nova objects

The VirtCPUModel is to present the guest cpu modelling. As
this is a quite new object, no from/to_dict function is needed.

The VirtCPUModel object has limitation to the arch and mode,
but not for model or features, so that no update needed when
new CPU model or features are released.

Co-authored: Daniel P. Berrange <berrange@redhat.com>
Change-Id: Ied869f587981804fa23f6b705b5da98a509790e2
Implements: blueprint resource-objects