Updated openstack/openstack Project: openstack/nova  005c100532e1a3a995874c28605d225565ab37b0
VMware: Enable spawn from OVA image

The VMware driver has not been honoring the container_format of "ova".
The change narrowly interprets any image with this container_format
value to be a single-file OVA image, and will attempt to stream the
first disk in the OVA as the root disk image.

The disk is guaranteed to be in streamOptimized format, which is
currently the only format we can deploy on regular and vSAN datastores.

Implements: blueprint vmware-driver-ova-support
Co-authored-by: Arnaud Legendre <alegendre@vmware.com>
Co-authored-by: Radoslav Gerganov <rgerganov@vmware.com>

Change-Id: Iddb7cca8af755f6cc0eb6861bce6cb29a4c18b50