Updated openstack/openstack Project: openstack/nova  d7d6f2cb34c35ad5a78c2f29189617d3a9817f20
Make tests use sha256 as openssl default digest algorithm

The tests previously used md5, which is considered broken,
and distros are starting to disable this in their openssl
builds.

Make the tests use sha256 as the default as a long term
sane alternative that should work on all distros.

This will fix Centos 7, and future proof the tests.

Closes-Bug: #1399498
Change-Id: Ic6cc92e47a318d789db3c3c98c67948eefb51fc2