Updated openstack/openstack Project: openstack-infra/puppet-project_config  32bba65a8a14054388ce04c403e59b84a6cefad5
Provide way to specify repo ref

We don't always want to use just 'master' since that may result in races
when running puppet across many nodes while project_config is changing.
Allow for a way to specify the ref we want via facter with a fallback to
master if the fact does not exist.

Change-Id: If1c17e55c7f1b15d28748761c3ee99af88fada40