Updated openstack/openstack Project: openstack/manila  24f8f7b6c91e98637ea0fec941ef96eccd524025
Manila cDOT qualified specs

Manila has "share types", analogous to Cinder "volume types", that
may contain qualified extra-specs - extra-specs that have significance
only for the backend driver and which are not reported back to the
scheduler.

This commit implements the following qualified extra-specs:
  * netapp:language
  * netapp:snapshot_policy
  * netapp:max_files
When creating the cDOT flexvol that backs a new Manila share, the flexvol
is provisioned with each of these characteristics if the Manila share is
of a share type that contains a corresponding qualified extra-spec that has
been set to the appropriate value of the extra_specs. If the value of a
particular extra_specs is not set, it defaults to a predefined value for
each of the extra_specs.

Partially Implements: blueprint cdot-qualified-specs

Change-Id: I1a70fe2d943abb4a48afb96d51f24227b0faa6e0