Updated openstack/openstack Project: openstack/nova  5675a2795bc73069120108b0bf91403498f82416
Ensure publisher_id is set correctly in notifications

We need to pass the service parameter else the publisher_id is
set to "None.node-1". There are many code paths that end up
in get_noftifier, so let's just handle it there just like we
default to CONF.host. Also the self.notifier in the
resource_tracker was not being used, so get rid of it.

Closes-Bug: #1416435
Change-Id: I8838b961e49496cc85913a04bd99f9845ff308c3