Updated openstack/openstack Project: openstack/keystone  8690c751abed5767c4dd0f5f41a0e08766f83cd8
Correct a v3 auth test for result ordering

The attribute picked by this test may not be one that causes an
Unauthorized exception to be raised since the order of attributes in
the list is random. This can be seen by setting the PYTHONHASHSEED
value to something other than 0.

With this fix, all the attributes are tested so the order doesn't
matter and the test coverage is improved.

Change-Id: I61f9efc5c60271c7ddf75eb4f4eedc66f6f4d67f
Partial-Bug: #1348818