Updated openstack/openstack Project: openstack/tempest  e0dd248d5150481769b062beaef0c915e4b53ed0
Add compute feature enabled flag for the ec2 api

The ec2 api's skip decision maker logic is convoluted and doesn't work
if we don't set any of the options. This commit adds an explicity config
flag to make it clear to say that your deployment doesn't support/use
the ec2 api and have the boto tests skipped.

Change-Id: I47f44830b1aa38c32682c6c3e6f619a3ca59b5f3