Updated openstack/openstack Project: openstack/python-keystoneclient  58ac2de5d4a6b58e8bd5d430a04199a4d40427a8
Reference identity plugins from __init__.py

Add the identity plugins to the __init__.py so they can be used without
specifically importing the v2, v3 or generic files.

This changes some usages of keystoneclient.discover to
keystoneclient._discover as the generic plugins are available at the
same time as other versioned plugins when keystoneclient.Client is
imported. This is the reason we have _discover in the first place.

Change-Id: I7b9bbc123aeac11d22b3a58395391d01af0427eb