Updated openstack/openstack Project: openstack/nova  713767660ba44421ea5e0406d121ea868ad178ac
network: Fix IPv6 tests for Mac

Some network tests would fail on a Mac because the IPv6 addressed returned
wasn't an exact match due to a difference in the `socket.inet_ntop` function
that `netaddr` uses.

On a Mac 0 parts of the IPv6 are compressed away to '::' whereas on Linux they
remain ':0:'.

Ultimately, the fix should be in the upstream `netaddr` library which already
has an issue logged for it: https://github.com/drkjam/netaddr/issues/82

Until that's fixed, the solution is to skip the offending tests.

Co-Authored-By: Johannes Erdfelt <johannes@erdfelt.com>
Change-Id: I7a8a49f4105079a686e553b4d002f04a26b93d0f
Closes-Bug: 1409135