Updated openstack/openstack Project: openstack/keystonemiddleware  fb306836b411546db385274996a8cdaf6154aaf7
Refactor auth_uri handling

The auth_uri setting is the value that's returned on auth failure
in the WWW-Authenticate header. It's not actually used by auth_token
to communicate with the identity server, so the IdentityServer
class doesn't need to know what the setting is for the auth_token
middleware. IdentityServer only needs to know how to provide the
default auth uri (based on how IdentityServer is configured) to the
auth_token middleware in case auth_uri isn't specified.

This change is moving things around to clarify how the auth_uri is
used.

Change-Id: I97aeef5721a044422929c72e9e53898822bcdabf