Updated openstack/openstack Project: openstack/keystonemiddleware  b562b04ee5db309268716e0e1b8270f30bdf1a76
Turn our auth plugin into a token interface

This auth plugin is already going to be distributed to everyone so that
they can use it to talk to use the current authentication to talk to
other services. Given that it knows everything that auth_token
middleware knows it may as well be the base for generating policy
contexts.

Implements: bp auth-context-object
Change-Id: I15e3517e854ef110987e392510c6de0f14645a59