Updated openstack/openstack Project: openstack/keystonemiddleware  02abaa1d2711a3d5fc0dd020f05133618e5b7dde
move add event creation logic to keystonemiddleware

currently, the logic to create audit events is contained in pyCADF.
this makes it extremely difficult to modify as changes often require
edits to two different libraries, and subsequent releases of two
different libraries. additionally, it makes it impossible to test the
full path. this patch attempts to minimise the chaos.

Change-Id: Ic0ca193139acaa751cd96e7fadad2223a84088c8