Updated openstack/openstack Project: openstack/oslo.utils  8a5a4d87da03919bf019a3c3ffeeddc3959b77d1
Add a eventlet utils helper module

When greenthreads are being used (and/or eventlet is available and
some modules have been monkey-patched) emit warnings to the users
that certain modules should be monkey-patched and if they are not
then spurious or unexpected lock-ups and/or hangs may result due to
this type of mixed usage which doesn't typically end well.

This commit adds that functionality to a new eventlet utils module
that can be used by calling code when that code really knows it will
not work without modules being patched or partially being patched.

Change-Id: I9a856fc0a6502b438c8da9b2f589154a6fa89a9f