Updated openstack/openstack Project: openstack/neutron  dd8804b06508b4471cc4e86d5d60fb8d39794a88
Remove use of keepalived 'vrrp_sync_group' as it is unused

Now keepalived configuration wraps the VRRP instances with a
'vrrp_sync_group'. The VRRP sync group functionality is only
relevant when more then one VR instance is contained in it.
In that case the VRs in the group will have the same state.
Our use of keepalived uses a single instance per router.

This patch simply removes the 'vrrp_sync_group'.
In this patch VR instances are used on their own and they now
hold the 'notify_scripts'.

Note that the same VRRP functionality is preserved with this
patch.

Another motiviation for this patch, aside from removing
useless configuration, is to lay the foundation for a future
patch that will the related bug by adding 'track_script'
that are not supported with 'vrrp_sync_group'.

Change-Id: I33b81049cd9cf140244bbf121d1a71492161c77c
Related-Bug: #1365461