Change scope from integrated gate to all of openstack
Instead of just tracking projects that are part of the integrated gate
lets start tracking all of openstack.

Defining openstack as everything in openstack/ openstack-infra/ and
openstack-dev/

As we move away from a single integrated gate, the list of what should
be in this repo under the old definition is quickly dropping. Instead of
using this to just help debug integrated gate failures, we can use this
to better track all openstack activity.

Knowing the basic nature of a gate failure means you can quickly narrow
down the number of repos that can possibly trigger the failure. That
combined with ElasticSearch data to identify a window of time the issue
was introduced during makes it easy to quickly narrow the number of
patches that could have possibly cause the problems.

Change-Id: I9c04a4d53f4628e0419d744b7e047ad75c0d133d