Updated openstack/openstack Project: openstack/nova  bb9964d059b007ea2e297638540141b497f71f64
VMware: update get_vm_resize_spec interface

This interface originally received the instance object as input.
This now receives the vcpus and memory_mb used by an instance. The
reason for the change is to enable us to either use a flavor or
and instance object to get the actual instance properties.

Change-Id: I789ffc06cea51e44150df8a50f51cbc10bf0f757