Updated openstack/openstack Project: openstack/horizon  fd56ce78c6b6c77103c90e6aea0caaba101cf19d
Set the password_autocomplete default to "off"

It's safer to set the autocomplete option to "off" for passwords
so that browsers get the hint to not save it. The default should
be secure so that deployers need to make a conscious decision to
be less-secure.

This is for security hardening.

SecurityImpact

Closes-Bug: 1420863

Change-Id: If2c3439cf23b11dd7829a4d7866d3b21409a7d69