Updated openstack/openstack Project: openstack/keystone  cb0018b79dd21b8b8c83421263c292d54c649fec
log wsgi requests at INFO level

Previously there was a random debug logging of the wsgi params,
disconnected from the requests coming in. This isn't useful to
anyone. Even though in an Apache env the access logs include the top
level requests, logging the wsgi requests, with params in the horizon
log is good form, and fits within the log guidelines.

Change-Id: I268b8b4b424f60c8bf4b3fb93bd8804c8f7bc65a