Updated openstack/openstack Project: openstack/neutron  368aaeaff40913343b78463216cb4f1b03feafcd
Refactor radvd control in the l3-agent

Several of parameters used by radvd are known when a router is created
and do not need to be passed around every time an RA method is
called. Also, we want to easily check the state of radvd for a router.

Use an object to keep track of the data and state of an radvd process.

Related-blueprint: restructure-l3-agent

Change-Id: I6bcad9d84af5a5b148df7520f582392a8b56a2ec