Updated openstack/openstack Project: openstack/nova  a913c5f470094c77d1708b3fc2363df0f03581d2
Add tests for nova-manage vm list

There were no existing unit tests for this nova-manage command and since
we just converted it to use objects we should probably make sure there
are unit tests for it.

Also fixes a little bug where instance.launch_index could technically be
None which would cause a NoneType error. The data model defines
launch_index as nullable but in practice the compute API will never
leave it as None, but we might as well fix this while adding tests.

Change-Id: Idc1ce6d11d7f6cba52cc6d8cd668b4386d1071ba