Updated openstack/openstack Project: openstack/neutron  86dbb886be84b5c0f6941bdf9f10c5965d0bd025
Move NCS mech driver to its new home

The NCS driver is scheduled for decomposition, and in the targetted vendor repo
the files are in a new location in the tree. Before we can decompose the driver
we must move it to its new location in the neutron tree.

A future patch will decompose (thin) the driver and add requirements.txt.

Partial-Bug: #1416713
Partially Implements: blueprint core-vendor-decomposition

Change-Id: If9d098bbbd571f53f945122e007d12d561295546