Updated openstack/openstack Project: openstack/taskflow  56edc108ab409d1648e8a715d2c6747869b83e08
Modify stop and add wait on conductor to prevent lockups

Removed _dead.wait from the stop method and added a wait method to
explicitly wait for graceful shutdown.  Implicitly waiting for
graceful shutdown in the stop method is leading to potential
deadlock situations when conductor is running in the main thread
and a signal handler is used to kick off stop/shutdown.

Also in larger systems, it is desirable to be able to initiate
shutdown on all running services then wait/join everything together
so as not to force a serial shutdown pattern.

Change-Id: Ic74dab22b09100c5b653d12c2518f83a1bf8e859