Updated openstack/openstack Project: openstack/keystone  bf050c6931bd3d72f1a03066c3c28a98e72d1529
Sync with oslo-incubator

This sync is crucial to adopting oslo.log, the bulk of the changes
are for the oslo modules to use python standard logging instead of
relying on oslo.log.

from keystone:
  rm -rf keystone/openstack/common

from oslo-incubator:
  python update.py ../keystone

Had to undo the changes to local.py and log.py since they are still
being used throughout keystone.

Also, with the sync came a change to versionutils, which now
emits a deprecation exception from versionutils instead of
the incubator's log.py. This meant that a test had to be updated
to catch the appropriate exception.

Sync'ed to level 03143669d2c996f7c20d2d75f6c67827d2db216a

Change-Id: If10eb5e2ea9c93dd22763f37faf465e103d9815d